tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/008.html

<!--
var log = postMessage;
importScripts('data:text/javascript,function run() { log(true) }');
run();
/*
-->
<!doctype html>
<title>variables and functions crossing importScripts boundary</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
<!--
*/
//-->
